Fix tests for logging connection info
While logging connection info in the glance cinder store driver, we
need to mask passwords for which we use oslo_utils's mask_dict_password
method.
Since this is not mocked in the glance functional tests, we encounter a failure
in cross-glance-tox-functional job in the change[1].
This patch adds the required mocking to pass the tests.

[1] https://review.opendev.org/c/openstack/glance_store/+/818722

Change-Id: Ibd9563fb854f888b4a2f5a5dcf6abce0cfab0f04
(cherry picked from commit acc6f624cdb1b2e4a3900e7d226bbc5156e8fa17)