[stable-only] Specify openstacksdk stable branch
Taking the advice given here:
http://lists.openstack.org/pipermail/openstack-discuss/2022-January/026905.html
to fix the openstacksdk-functional-devstack job.

This job is not specified in glance's .zuul.yaml, but is brought in
by the integrated-gate-storage template:
https://opendev.org/openstack/tempest/src/commit/86db21ea6afb6c26a21fc143a7d061e947c31a93/zuul.d/integrated-gate.yaml#L382-L398

So we add the job here so that we can override its definition.

Change-Id: I8f1f38df4746cced085fb1cb4e96540763cdda13
(cherry picked from commit b76dd7880cf40f8f2cc9425c54941c5d637401a2)