Fix warning in the release notes
Glance yoga release notes generates one warning and as a result
it will fail to list it on release page [1] like it is
showing for Xena release.

The warning in the test log is;
"The Fixes section of
releasenotes/notes/fix-md-tag-create-multiple-c04756cf5155983d.yaml is
not a recognized section. It should be one of: features, issues,
upgrade, deprecations, critical, security, fixes, other. This
will be an error in a future release."

Rmoved the "Fixes" section from mentioned release note as it is already
included in `yoga-rc1-release-notes-153932161f52a038.yaml`.

[1] https://releases.openstack.org/teams/glance.html#team-yoga-glance

Closes-Bug: #1963282
Change-Id: I0167cd65f37fc744aa686c626739f9563054666c