[CI] Add upper constraints to install command
In cases when a package is installed in tox's develop-inst phase upper
constraints are not used if it were not defined in 'deps' of the tox
target. This caused for example 'tox -e docs' to fail in case a package
introduced some branch incompatible change.

There are two option to fix this:
* add the upper constraints to 'deps' of the target (and pre-install
  requirements for some cases) OR
* use it in the install_command

In the past the latter needed some extra attention (due to e.g. lower
constraints tests) and projects were encouraged to remove the
install_command, but nowadays this seems better fitting for some
project and easier to use. So this patch fixes the upper constraints
usage via the custom install_command and by removing it from 'deps'.

Conflicts:
  tox.ini

NOTE(elod.illes): conflict is due to branch specific upper constraints
link and variable name change (UPPER_* to TOX_*).

Change-Id: Iae07e1c5abde09874928951ac55c2d0355bcd244
(cherry picked from commit 36aedba20e998953333225a9788c084d2e6be6ae)