Disable import workflow in glance cinder jobs
Recently, glance-multistore-cinder-import job started failing.
As per the RCA done here[1], the reason is glance is using
import workflow to create images which is an async operation.
As in case of glance cinder configuration, there are a lot of
external APIs (cinder) called like volume create, attachment
create, attachment update, attachment delete etc which takes
time to process hence the image doesn't get available in the
expected time (as per devstack) hence the failure.
Disabling import workflow will cause the images to be created
synchronously which should pass the glance cinder jobs.
To disable import workflow, we are inheriting from
tempest-integrated-storage and not
tempest-integrated-storage-import (which has import plugin enabled).

[1] https://review.opendev.org/c/openstack/glance/+/841278/1#message-456096e48b28e5b866deb8bf53e9258ee08219a0

Closes-Bug: 1973136
Change-Id: I524dfeb05c078773aa77020d4a6a9991a7eb75c2