Fix failing namespace list delete race
If a namespace is deleted by another client while we are doing a
namespace list operation, we will fail the list with NotFound if we
try to pull the resource_type_associations list. The latter re-queries
the DB for the namespace and will raise NotFound to us. This is
especially bad because the namespace being deleted need not even
belong to the caller of the list, as is the case in a tempest run.

This makes us catch the failure and continue the operation, reporting
no associations so that the client gets a consistent view and no
error.

Closes-Bug: #1973631
Change-Id: I09fc9164a08f42507d2aec44c5b382a72f232571
(cherry picked from commit 35e6c57bc90365796f45ada9271e532068557b8a)
(cherry picked from commit cc98dbef31be30372b59b2d18b78694ce0879aa0)