Include os-brick configuration options
The cinder glance store now supports using an os-brick specific location
for file locks instead of using the one defined for the whole service
using the ``lock_path`` configuration option in the ``os_brick``
configuration section.

To render this option when generating the sample config file with ``tox
-egenconfig``, the ``os_brick`` namespace needs to be configured in the
oslo-config-generator config files and the os-brick dependency needs to
be present.

This patch adds the os-brick package as a dependency in the genconfig
test environment and adds the namespace to the relevant config files.

Depends-On: Ib11d63e3c388e12f145f40247d17030a566b8c5e
Change-Id: Ie25ab8bcc01a57e26da6a951e899659ec1782140