Mock getaddrinfo in cinder store tests
With change Ic79815972e654a8bfe2775f57c68cfa0bf115e2f, we started
using getaddrinfo method from socket lib to fetch the system
IP address (v4 or v6) and pass it to os-brick.
The test works for most systems since a valid IP is returned but
for some deployments, it fails with the following error.

raise socket.gaierror(socket.EAI_NONAME, 'No address found')

In any case, it is a good practice to mock it in tests.
This patch mocks getaddrinfo in the current cinder store tests.

Change-Id: Ib03fefbb6034a7e7d2ff1e16430d28962c2ca355