Fix unintentional exception inspecting VMDK
It looks like a raise statement was left in the virtual_size property
handler for VMDK, which should have been converted to a log at some
point. All the other inspectors return zero for virtual_size if the
format does not match or they are unable to parse the data. This
converts that raise to a log, and adds a test to make sure we make it
far enough in the processing of the complex VMDK format to ensure that
behavior.

Closes-Bug: #1983279

Change-Id: I0352ab6b2c00055de094ac5902b8d50941d06dcf
(cherry picked from commit 199722a65a88f9c28b554b180ec36415ed3a51ac)
(cherry picked from commit 07f98747a2d4bf65f44348603ecba639d602db4d)