Fix cinder tests for refactoring effort
We are refactoring the glance cinder store[1] to support the
extend in-use feature[2] due to which cinder store tests on
glance side are failing[3] as they aren't able to locate the
files in their new location.
This patch fixes the issue to import from new path while keeping
backward compatibility to import from old path to pass gate until
the new changes have merged.

[1] https://review.opendev.org/c/openstack/glance_store/+/843103
[2] https://review.opendev.org/c/openstack/glance_store/+/868742
[3] https://zuul.opendev.org/t/openstack/build/41a1abf9c749476bb10a56600ca07f0e

Change-Id: Ib9f7160fdbac74fc419faa6b3ab8acb17400d392