Fix functional test failures with PasteDeploy 3.x
PasteDeploy replaced pkg_resources with importlib_metadata[1] in 3.0.
This causes functional test failures because the line breaks in the
example configs from glance/tests/functional/__init__.py are read as a
"prefix" by PasteDeploy. For instance,

paste.filter_factory =
  glance.api.middleware.version_negotiation:VersionNegotiationFilter.factory

is read as "\nglance.api.middleware.version_negotiation:VersionNegotiationFilter.factory".

This might be a bug in PasteDeploy or (quite unlikely) in
configparser.ConfigParser rather than in our code, so this issue should
be revisited at some point. For now, remove the cosmetic line breaks in
order to fix our gates.

[1] https://github.com/Pylons/pastedeploy/commit/ef58c20961072e4f60c26e94564180c304af7144

Change-Id: I1e557e731cceb3cfbc4ebc3f0887adf372a91d1d