Update functional jobs for 2023.2
Based on the python runtimes for 2023.2 [0], we should be running
functional jobs on python 3.9 and python 3.10.  This patch adds
functional testenvs for python 3.10 and 3.11 (the latter for local
testing) to tox.ini so the appropriate zuul jobs can be defined.

The functional-py38 testenv is not removed from tox.ini, as it may
be useful locally.

Functional py39 and py310 jobs are added to the check and the gate.

The current openstack-functional-py38-fips job is not removed; a
py39 job has not yet been defined to replace it [1].

[0] https://governance.openstack.org/tc/reference/runtimes/2023.2.html
[1] https://opendev.org/openstack/openstack-zuul-jobs/src/commit/b3cad4f7a36929c13d9c74f2de24789741aa5d85/zuul.d/jobs.yaml

Change-Id: Ibc21b107878f5ba50137da4082a7cbc6342d2aa9