Test glance RBAC old defaults
Glance has enabled the scope checks and new defaults in
antelope release and now devstack is switching the testing
to new defaults by default (depends-on). With that all the
jobs will run the the new defaults.

As old defaults are still supported (as deprecated), we should
test those at least in a single integrated job. Enable the old
defaults in glance-multistore-cinder-import job.

Keeping existing job enabling the new defaults because other services
have not switched to new defaults yet to this job enable new defaults
for those service and test Glance against that.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/883601/
Change-Id: I470e3b8c1106d88e85343508a8e5891c88861c98