Handle case for multiple cirros images
The change[1] configures multiple images for the devstack-tempest
job and all the jobs deriving from it. This causes problem with
the current post-check-metadata-injection.yaml playbook since
we only expect one cirros image and perform steps accordingly.
This change modifies the playbook to handle multiple cirros
images.
Failure in the glance-multistore-cinder-import job can be
seen here[2].

[1] https://review.opendev.org/c/openstack/tempest/+/831018
[2] https://zuul.opendev.org/t/openstack/build/68a4d3ec6ce04c87b21d73a333f5b5cd/log/job-output.txt#23161

Change-Id: I70b4a970d7e16174f715c54bdae9467cc321c13e