Correct item numbering in install docs
Indicating a new enumerated point by a '#.' in an included doc
restarts the numbering at 1 instead of continuing the current
sequence.  Move the '#.' sentence out of the included doc and
into the individual install docs so that the points are enumerated
correctly.  Also remove one level of indentation from the included
file and instead indent the 'include' directive.

(Also had to break two lines in register-quotas.rst due to doc8
D001.  Not sure why that suddenly broke now after I shortened
all the lines in that file!)

Change-Id: I53c136c728fcffbca07eb5d0a67f9edc5e723ad0