db: Replace use of aliased keyword arguments
Resolve the following RemovedIn20Warning warning:

  The ``aliased`` and ``from_joinpoint`` keyword arguments to
  Query.join() are deprecated and will be removed in SQLAlchemy 2.0.

This one is a little trickier to grok, but the SQLAlchemy docs explain
it rather well [1] (Look for "Legacy Features of Query.join()" and
"Joins to a Target with an ON Clause"). Essentially, because we are
joining e.g. the ImageTag table multiple times, we need to create an
alias and use it for both the join and resulting queries.

[1] https://docs.sqlalchemy.org/en/14/orm/query.html#sqlalchemy.orm.Query.join

Change-Id: Ifb818df8ca8500adcf1354eff0671f5762473110
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>