Fix post failures causing by image conversion failure
CI jobs tempest-integrated-storage-import-standalone,
tempest-integrated-storage-import and
glance-multistore-cinder-import were failing while converting
image from qcow2 to raw. The failures were caused by changing
default behavior of Devstack and installing everything in global
virtualenv patch [1].

This fixes import conversion in the affected jobs properly by setting
the python interpreter to the one in the venv. When we spawn external
commands they actually spawn a python interpreter to use an oslo
module to do the exec, and we need to do that with our venv's python.

Finally, this bumps the swap size for the storage-import jobs to 8G
to help account for the additional overhead of running another glance
worker, which is required for that configuration. This matches other
jobs with additional overhead (i.e. the ceph ones).

[1] https://review.opendev.org/c/openstack/devstack/+/891248

Co-Authored-By: Dan Smith <dansmith@redhat.com>
Depends-On: https://review.opendev.org/c/openstack/devstack/+/892571
Change-Id: I073216d1bbddbd3bdecd2b135e4c291f7a4d9abe