Increase timeout for glance cinder multistore job
Currently the glance cinder multistore job times out from time to
time in 2 hours (and some minutes)[1].
The issue is not specific and generally points to a slow devstack
deployment or a slow tempest run[2] which exceeds the allocated time
provided for the job to complete.
Unless the job is checked thoroughly to determine the areas of
improvement, a short term fix would be to increase the timeout
of job (which might also be the ideal solution if the job is
landing on a system with inferior hardware).
This patch increases the timeout to 10800 seconds i.e. 3 hours.

[1] https://zuul.openstack.org/builds?job_name=glance-multistore-cinder-import&project=openstack%2Fglance&skip=0
[2] https://review.opendev.org/c/openstack/glance/+/899868/1#message-081c737d8747de44669acd1cdc322e347b09d390

Change-Id: I2d21d8b098355a6cc4ceb1de8f9d7fefa19be965