Fix flaky test related to cache migration
In cache migration unit tests, I was using same cache.db file name
in the setup method which was used by all the tests, which was
causing a issue of deleting a file by other test and resulted
in random failure.

Using random filename to fix this issue permenantly.

Related blueprint centralized-cache-db

Change-Id: I321982013bdf571741a8ef2d0575f65288fff53f