Set a lock_path for tests
oslo.concurrency 6.0.0 has relocated the lock_path that is required
for external locks from [DEFAULT] to the [oslo_concurrency] config
section.  Instead of messing with configuration, we can use the
lockutils.set_defaults() method to set this for tests.  So add this
to the BaseTestCase so that it's available for all tests.

Change-Id: I2878b4162dddaf935002c168bc544eb1eff71efd