Cinder: Remove ambiguous warning
In case of Cinder store, when updating location of a legacy image,
we call cinder store to perform the update.
If the operation fails, we log a warning in glance which states:

"Not able to update location url <url> of legacy image due to unknown issues."

Since glance doesn't have knowledge about the actual issue occurred
during the location update, we log this ambiguous warning which is not
useful for operators to debug the actual issue.

The responsibility of logging the warning should be delegated to
glance store (like done here[1]) since it can point to the right
issue as it knows which operation failed exactly.

This patch removes the warning and further warnings should be added
to glance store only.

[1] https://review.opendev.org/c/openstack/glance_store/+/911005

Change-Id: I24b14812a2a15675beacee720d722c2f3a2b252b