Increase timeout for tempest-integrated-storage-import job
The tempest-integrated-storage-import has been timing out
frequently around 2 hour 5 minutes mark and similarly one
of the jobs inheriting from it,
tempest-integrated-storage-import-standalone, also times
out around the same time. The timeout issue has increased
recently as can be seen in the job build history[1].

We had a similar issue with the glance-multistore-cinder-import
job for which we increased the timeout to 3 hours[2] but the
job completes successfully on average 2 hour 10 minutes
and has been completing in that time for more than 8 months
since the change to increase timout landed.

The underlying cause could be due to a lot of reasons:
1. Lower concurrency (3) as compared to parent job (6)
2. More number of tests being added to tempest
3. Slow Devstack installation
4. Slow tempest test execution
5. memory consumption of OpenStack services
increased making environment slow
6. no significant improvement in the hardware
specifications on which we run the job

This patch generalizes the timeout to the tempest-integrated-storage-import
job so all of it's child jobs also don't face similar timeouts.

Also we are updating timeout for tempest-integrated-storage-enforce-scope-new-defaults
job as we see similar time out issues in that job.

[1] https://zuul.opendev.org/t/openstack/builds?job_name=tempest-integrated-storage-import&skip=0
[2] https://github.com/openstack/glance/commit/64e6da6fe1c810befab1e5a59230ce8711dce709

Change-Id: I7165bb8c588659362b7b36394111c41772ac802d