Make glance able to require disk_format match
This adds a config knob to enable strict format compliance when
uploading image data. If enabled, uploads will be rejected for formats
we can inspect if the content does not match. For formats we cannot
match (other than raw), we will reject content if we _do_ recognize
the format.

Related to blueprint glance-as-defender
Depends-On: https://review.opendev.org/c/openstack/tempest/+/931028
Depends-On: https://review.opendev.org/c/openstack/glance-tempest-plugin/+/931287
Depends-On: https://review.opendev.org/c/openstack/nova/+/930754
Change-Id: I2d0301d168d2b55bf38351bf5c6d780e95a00401