Doc: Add info about multi-attach volume types
When using Cinder as Glance backend, we can run into an issue
when using the same image to create multiple resources.
Example: creating 10 bootable volumes from an image

This happens because, by default, Image-Volumes (storing the
image) does not allow to be attached multiple times causing
the issue to not being able to use the same image concurrently.

This patch adds a doc section to configure multi-attach volume
types to be used for Image-Volumes to prevent the above issue.

Change-Id: Idcfaffe0a3f1a030ccffd6a5fc5e7b6523ae015a