Fix performance glitch while sorting image locations
Some of the available glance stores like file, cinder etc has
capability to reuse already initiated driver (DRIVER_REUSABLE = 0b01000000). In
Caracal we have added a feature to sort image locations based on store
weight. As RBD driver of glance does not have this reuse capability, during
image list API call it initializes the RBD driver for each of the available
image which is causing noticable delay in list call.

To avoid this, using new interface added in glance_store which will directly
get the weight of the store from memory and return it back to user.

Depends-On: https://review.opendev.org/c/openstack/glance_store/+/934362
Closes-Bug: #2086675
Change-Id: I662ba19697e03917ca999920ea7be93a0b2a8296