Add ``fsid`` to RBD store properties
Currently the RBD store/backend doesn't report the FSID field.
To include the FSID field, we need a connection to the RBD cluster.
Luckily, while initializing the RBD store, we fetch the FSID and
put it in the _url_prefix field (given fsid and pool info are available).

We just need to fetch it from the _url_prefix field and return it
in the response.
If FSID is not set in the RBD store, we return NULL value.

This is required for the effort of optimizing the upload volume to
image operation from cinder RBD to glance RBD backend.

Partial-Implements: blueprint optimize-upload-volume-to-rbd-store
Change-Id: I3e7dc11bc632ca0f3e134e0bca7eb2442bf797ca