Install qemu-utils when doing unit tests
"test" profile is missing on qemu-utils to make sure the worker will
actually have qemu-utils installed, we need to force this for dpkg
platforms.

Note that it seems opendev workers are installing qemu-utils by default,
but this is maybe not always true in other CI environments.

Enforcing this is nice for consistency and reproductibility.

Change-Id: I6c3d763e920f0a4c3ce33429d21a122fb634ad74
Signed-off-by: Arnaud Morin <arnaud.morin@ovhcloud.com>