Remove FunctionalInitWrapper
This is part of our effort to move away from FunctionalTest. The
FunctionalInitWrapper class was only used in
glance/tests/functional/db/test_sqlalchemy.py.

The only thing FunctionalInitWrapper did was set the policy_file in the
fake config; this is also done through the set_policy() method of the
BaseTestCase class, so it is no longer needed.

Change-Id: Ia8e8d51fbea5b0e0edb3863a235ddc52af9a4d65