Functional tests migration: test_sqlite.py
Use SynchronousAPIBase instead of FunctionalTest as the base class for
the TestSqlite class, so we can run the tests it defines without running
eventlet or running a "real" WSGI server.

This also fixes this test which had not been doing anything for a while.
Because the database file "tests.sqlite" did not exist (it should have
been "{self.test_dir}/tests.sqlite"), the returned schema was empty.
Obviously, 'BIGINT' could not be found in that, so the test passed.

Change-Id: I2cae99af61196388ec842bb0d7f4c6c065ac18e8