Allow ignoring specific gpt safety checks
This allows the operator to configure specific GPT safety checks to be
non-fatal. This could be necessary if an operator finds that legit,
but non-spec-compliant images are already present in the system and
needs time to fix them. See this nova bug for an example of a widely-
distributed image that violates the GPT spec for example:

https://bugs.launchpad.net/nova/+bug/2091114

For maximum compatibility, we ignore the mbr check by default for the
first release with this check.

Related to blueprint glance-as-defender

Change-Id: Ibf9f5eb4e2378ddd77ebe61170b8e921ada792d7