Safety check output of image conversion
This makes us check the format of the resulting image after conversion,
both for format compliance and safety. Because it would be possible
to upload a complex format (like vmdk or qcow2) that unfolds into
another such format when converted to raw, we should make sure that
the result of that process yields something we expect.

This was originally part of change:

 I32416d38e295d32ae24652ec83716ca82eb76b06

but was pulled out to stand alone for earlier merging.

Related to bp/glance-as-defender

Change-Id: I53403d29ce160b34f1258ba2d3b28163de799593