Cancel hashing operation if image is deleted
Used file based locking to track the progress of hash calculation
task. If image is deleted in between and delete call is not
processed on the host where hash operation is in progress then
it will redirect to that host, cancels the hash operation and
then deletes the image.

Closes-Bug: #2045769
Change-Id: I956804491170b92686dc4de7b924371356b2626a
(cherry picked from commit f84e4f0123306a9bb47846cd73fc78578e7825aa)
Signed-off-by: Francesco Pantano <fpantano@redhat.com>