Handle images detected as ISO+GPT
ISO images can have a system-specific bootloader prepended to them,
which in the case of x86-64 would be a typical GPT/MBR boot sector,
partition table, etc. These files are not uncommon amongst installer
media and currently trigger and fail our format detection since they
are identified as both ISO and GPT. This makes us allow these files
if they are being uploaded with disk_format=iso as the target. No
other multi-format combinations are allowed.

Note that this only impacts the upload method and not import with
format conversion. The latter is more complicated as we currently use
oslo's detect_file_format() method which specifically disallows
multiple formats. Converting from iso+gpt to anything else is not
likely very useful and we already ignore it in the process (although
after the inspection check).

Closes-Bug: #2127789
Change-Id: I11a6872275568e874236ae78faf877db1bbfbb9a
Signed-off-by: Dan Smith <dansmith@redhat.com>