Migrate py313 job to Debian Trixie
Python 3.13 is the default version in Debian Trixie so we can use
more efficient package provided by the distribution.

Note that this also limits installation of PCRE(3) in doc target
because it is required by whereto, which is only called in the doc
target. We need this change because PCRE3 is no longer available in
Debian Trixie[1].

[1] https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/message/KVXZIET5ZK3C6I7S45O4WWYPGBVFDIWC/

Change-Id: I9d7a48055a7ebb7729267ffda92a7dd61ef64e3d
Signed-off-by: Takashi Kajinami <kajinamit@oss.nttdata.com>